backend/requirements.txt
# FastAPI and core dependencies
fastapi[standard]==0.115.12
uvicorn[standard]==0.32.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.9
alembic==1.14.0

# Authentication and Security  
PyJWT==2.8.0
cryptography==45.0.4
bcrypt==4.1.2
passlib[bcrypt]==1.7.4
python-multipart>=0.0.18

# Configuration
pydantic==2.10.4
pydantic-settings==2.7.0

# OpenAI
openai==1.58.1

# Development and Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
